Adicionando validações

+++ src/main/java/com/example/desafio5/DadosNumeros.java
@@ -1,4 +1,4 @@
 package com.example.desafio5;
 
-public record DadosNumeros(int[] numero) {
+public record DadosNumeros(double[] numero) {
 }
